Multiplayer Rock Paper Scissors Android App
This application uses voice as the input for the game. The application
can be played in single player mode against a bot or can be played by
two users over bluetooth. Application stores the win and lose data in
the local database.
Future enhancements would be to connect the application between
multiple users over Wifi and to store the data over cloud
infrastructure.